[Encoding] Allow `PadEncodingAttribute` to support dynamic padding. (#20662)
The dynamic padding allows different backends to chose different padding
values.

To be able to use `?` for representing dynamic values, the container is
changed to be int64_t.

+++ compiler/src/iree/compiler/Dialect/Encoding/IR/unittests/EncodingAttrTest.cpp
@@ -61,7 +61,7 @@ TEST_F(EncodingAttrsTest, PadEncodingLayoutAttr) {
   EXPECT_TRUE(cast<SerializableEncodingAttrInterface>(zeroPaddingAttr)
                   .isIdentityLayout());
 
-  SmallVector<int32_t> paddings = {4, 2};
+  SmallVector<int64_t> paddings = {4, 2};
   auto nonZeroPaddingAttr = PadEncodingLayoutAttr::get(ctx, paddings);
   EXPECT_FALSE(cast<SerializableEncodingAttrInterface>(nonZeroPaddingAttr)
                    .isIdentityLayout());
